examples/web/pyproject.toml
[project]
name = "web-search-examples"
version = "0.1.0"
description = "Web search cookbook examples for SmartSearch SDK"
requires-python = ">=3.11"
dependencies = [
    "smart-search-sdk",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]
